Pull in the Multi Instance branch.  Allows installing multiple instances of built-in apps. (#1874)
* Replace all app names with variables in templates

* Change template app name variables to `<appname>` and `<APPNAME>`

* Revert "Change template app name variables to `<appname>` and `<APPNAME>`"

This reverts commit 41bd2cd516c69fa7f06f9185ba325fb0421fff0d.

* Revert "Merge pull request #1788 from CLHatch/multi-instance"

This reverts commit e615f6bf8c1907155e3735dd3d541f565410bf94, reversing
changes made to 5737b6bc21e4f041089f549857b319539c1a3a69.

* Update update_self.sh

* Add `<__INSTANCE>`, `<__Instance>` and `<__instance>` tags to template files

* Update appname_is_valid.sh

Don't allow certain instance names.

* Create instance_file.sh

* Update update_self.sh

Clear out the instances folder when updating DockSTARTer.

* Rename `.env` to `appname.global.env` and `appname.env` to `appname.app.env`

* Update instance_file.sh

* Create instance_folder.sh

* Update update_self.sh

* Update instance_folder.sh

* Update instance_file.sh

* Update instance_file.sh

* Update instance_folder.sh

* Update instance_file.sh

* Update instance_folder.sh

* Update instance_folder.sh

* Update instance_file.sh

* Fix double `<__Instance>` in hostname vars

* Update instance_file.sh

Always return he file path, even if the file doesn't exist.

* Update instance_folder.sh

Always return he folder path, even if the folder doesn't exist.

* Update instance_folder.sh

Remove unused variable

* Update instance_file.sh

Remove unused variable

* Update instance_file.sh

* Use `instance_file` and `instance_folder` to access all templates

* Update app_is_nondepreciated.sh

Remove unused variable.

* Update menu_app_select.sh

Remove unused variable.

* Update app_is_builtin.sh

* Update app_description.sh

* Update menu_app_select.sh

* Update app_list_enabled.sh

Use `instance_folder`

* Update app_is_builtin.sh

* Update appvars_create.sh

* Update app_is_builtin.sh

* Update appvars_create.sh

* Update app_is_builtin.sh

* Update appname_is_valid.sh

* Update appvars_create.sh

* Update appvars_create.sh

* Update yml_merge.sh

For depreciated apps, give a warning, but still allow them to run.

* Rename `instance_file` to `app_instance_file`, and `instance_folder` to `app_instance_folder`

* Update lint.yml

Disable MD059/descriptive-link-text Link text should be descriptive [Context: "[here]"]

* Create an `app_templates_file` function, and have `app_instance_file` use it.

* Update menu_app_select.sh

No need to check for or even know where the instance folder is here.

* Update menu_value_prompt.sh

Use the `app_instance_file` function to access the template file.

* Update app_instance_file.sh

Remove warning for template file not existing.

* Update menu_value_prompt.sh

Re-order the restart options to match the order in the description

* Create app_env_file.sh

* Update app_env_file.sh

Check for missing `env_files` folder when getting `appvars.env` file path.

* Usw `app_env_file` to get the path to an app's `appname.env` file

* Adjust some notices as bit

* Update main.sh

Declare global dialog colors and default dialog box timeout for dialog boxes that timeout.

* Use the global dialog color variables

* Update main.sh

Reset to "No Color" before all the menu default color codes.

* Update main.sh

* Update menu_app_select.sh

* Update main.sh

* Rename `DialogTimeout` to `DIALOGTIMEOUT`

* Update main.sh

Use the global dialog color variables.

* Remove `Color` from the pre-defined dialog color names

* Make a global "User Defined" heading color

* Update varname_is_valid.sh

* Update menu_add_var.sh

Fix checking for valid variable names.

* Change `*User Defined*` to `(User Defined)`

* Update menu_value_prompt.sh

Show if an app is depreciated or disabled.

* Update menu_value_prompt.sh

Show if an app is depreciated or disabled.

* Fix `app_is_enabled` and `app_is_disabled` for the case when the app is not built-in

* Update menu_value_prompt.sh

* Rename the `HeadingUserDefined` color to `HeadingTag`

* Update menu_add_var.sh

Add User Defined, Depreciated and Disabled tags to the application heading line.

* Update menu_app_vars.sh

Add User Defined, Depreciated and Disabled tags to the application heading line.

* Change `*DEPRECIATED*` to `[*DEPRECIATED*]` to match the `.env` file headings.

* Update menu_app_vars.sh

* Create app_is_user_defined.sh

Returns if an app ais user defined.

An app is considered "User Defined" if there is no `APPNAME__ENABLED` variable or the app is not built in.

* Create app_is_referenced.sh

* Update app_is_referenced.sh

* Update app_description.sh

Use `app_is+user_defined`

* Update app_nicename.sh

If the app is user defined, don't use the "nicename" from the labels file even if it exists.

* Revert "Update app_nicename.sh"

This reverts commit 16b94ad400f9d3e43e505e4601de178493e6a9e5.

* Revert "Update app_description.sh"

This reverts commit 970d545740426aa6dd02a4804b538184a0b857f2.

* Update menu_app_vars.sh

* Update menu_app_vars.sh

* Update app_is_user_defined.sh

* Update app_is_user_defined.sh

* Update menu_value_prompt.sh

* Update menu_add_var.sh

* Update app_is_user_defined.sh

* Reapply "Update app_description.sh"

This reverts commit d129ef74de000a5235ab5430452011755c17e1dd.

* Reapply "Update app_nicename.sh"

This reverts commit 99a565f61485eb7e9406d1657f464a752cf89a8d.

* Create app_description_from_template.sh

* Create app_nicename_from_template.sh

* Update menu_app_select.sh

* Update app_is_user_defined.sh

* Update app_description_from_template.sh

* Update app_nicename_from_template.sh

* Update some tests

* Update env_format_lines.sh

Change variable names

* Update env_format_lines.sh

Use variables for the 'Depreciated', 'Disabled' and 'User Defined" heading tags.

* Update env_update.sh

Check if app is not user defined instead of if it is added.

* Update env_format_lines.sh

* Update env_format_lines.sh

* Update env_format_lines.sh

* Update menu_value_prompt.sh

* Update env_format_lines.sh

* Update menu_value_prompt.sh

* Update env_format_lines.sh

+++ .scripts/menu_app_vars.sh
@@ -3,34 +3,6 @@ set -Eeuo pipefail
 IFS=$'\n\t'
 
 menu_app_vars() {
-    # Dialog color codes to be used in the GUI menu
-    # shellcheck disable=SC2168 # local is only valid in functions
-    local \
-        ColorHeading \
-        ColorHeadingValue \
-        ColorHighlight
-    # shellcheck disable=SC2034 # variable appears unused. Verify it or export it.
-    {
-        ColorHeading='\Zr'
-        ColorHeadingValue='\Zb\Zr'
-        ColorHighlight='\Z3\Zb'
-    }
-    # shellcheck disable=SC2168 # local is only valid in functions
-    local \
-        ColorLineHeading \
-        ColorLineComment \
-        ColorLineOther \
-        ColorLineVar \
-        ColorLineAddVariable
-    # shellcheck disable=SC2034 # variable appears unused. Verify it or export it.
-    {
-        ColorLineHeading='\Zn'
-        ColorLineComment='\Z0\Zb\Zr'
-        ColorLineOther="${ColorLineComment}"
-        ColorLineVar='\Z0\ZB\Zr'
-        ColorLineAddVariable="${ColorLineVar}"
-    }
-
     local APPNAME=${1-}
     APPNAME=${APPNAME^^}
     local appname=${APPNAME,,}
@@ -38,28 +10,40 @@ menu_app_vars() {
     AppName=$(run_script 'app_nicename' "${APPNAME}")
     local Title="Edit Application Variables"
 
-    if ! run_script 'app_is_builtin'; then
-        local Message="Application '${AppName}' does not exist."
-        if [[ ${CI-} == true ]]; then
-            warn "${Message}"
-        else
-            dialog --title "${Title}" --msgbox "${Message}" 0 0
-        fi
-        return
-    fi
-
     local AddVariableText='<ADD VARIABLE>'
 
-    run_script_dialog "${Title}" "Creating variables for ${AppName}" 1 \
-        'appvars_create' "${APPNAME}"
-
-    local DefaultGlobalEnvFile="${TEMPLATES_FOLDER}/${appname}/.env"
-    local CurrentGlobalEnvFile
+    local CurrentGlobalEnvFile CurrentAppEnvFile
     CurrentGlobalEnvFile=$(mktemp)
-
-    local DefaultAppEnvFile="${TEMPLATES_FOLDER}/${appname}/${appname}.env"
-    local CurrentAppEnvFile
     CurrentAppEnvFile=$(mktemp)
+
+    local DefaultGlobalEnvFile=''
+    local DefaultAppEnvFile=''
+    local AppIsUserDefined=''
+    local AppIsDisabled=''
+    local AppIsDepreciated=''
+
+    if run_script 'app_is_user_defined' "${appname}"; then
+        AppIsUserDefined='Y'
+    else
+        DefaultGlobalEnvFile="$(run_script 'app_instance_file' "${appname}" ".global.env")"
+        DefaultAppEnvFile="$(run_script 'app_instance_file' "${appname}" ".app.env")"
+        if run_script 'app_is_disabled' "${appname}"; then
+            AppIsDisabled='Y'
+        fi
+        if run_script 'app_is_depreciated' "${appname}"; then
+            AppIsDepreciated='Y'
+        fi
+    fi
+
+    local AppNameHeading="Application: ${DC[Heading]}${AppName}${DC[NC]}"
+    if [[ ${AppIsUserDefined} == 'Y' ]]; then
+        AppNameHeading="${AppNameHeading} ${DC[HeadingTag]}(User Defined)${DC[NC]}"
+    elif [[ ${AppIsDepreciated} == 'Y' ]]; then
+        AppNameHeading="${AppNameHeading} ${DC[HeadingTag]}[*DEPRECIATED*]${DC[NC]}"
+    fi
+    if [[ ${AppIsDisabled} == 'Y' ]]; then
+        AppNameHeading="${AppNameHeading} ${DC[HeadingTag]}(Disabled)${DC[NC]}"
+    fi
 
     local LastLineChoice=""
     while true; do
@@ -70,7 +54,7 @@ menu_app_vars() {
         local -i LineNumber=0
         local FirstVarLine
         ((++LineNumber))
-        LineColor[LineNumber]="${ColorLineHeading}"
+        LineColor[LineNumber]="${DC[LineHeading]}"
         CurrentValueOnLine[LineNumber]="*** ${COMPOSE_ENV} ***"
         run_script 'appvars_lines' "${appname}" > "${CurrentGlobalEnvFile}"
         local -a CurrentGlobalEnvLines
@@ -84,30 +68,30 @@ menu_app_vars() {
             VarName="$(grep -o -P '^\w+(?=)' <<< "${line}")"
             if [[ -n ${VarName-} ]]; then
                 # Line contains a variable
-                LineColor[LineNumber]="${ColorLineVar}"
+                LineColor[LineNumber]="${DC[LineVar]}"
                 VarNameOnLine[LineNumber]="${VarName}"
                 if [[ -z ${FirstVarLine-} ]]; then
                     FirstVarLine=${LineNumber}
                 fi
             elif (grep -q -P '^\s*#' <<< "${line}"); then
                 # Line is a comment
-                LineColor[LineNumber]="${ColorLineComment}"
+                LineColor[LineNumber]="${DC[LineComment]}"
             else
                 # Line is an unknowwn line
-                LineColor[LineNumber]="${ColorLineOther}"
+                LineColor[LineNumber]="${DC[LineAddVariable]}"
             fi
         done
         ((LineNumber++))
         local AddGlobalVariableLineNumber=${LineNumber}
         CurrentValueOnLine[LineNumber]="${AddVariableText}"
-        LineColor[LineNumber]="${ColorLineAddVariable}"
+        LineColor[LineNumber]="${DC[LineAddVariable]}"
         ((++LineNumber))
         CurrentValueOnLine[LineNumber]=""
-        LineColor[LineNumber]="${ColorLineOther}"
+        LineColor[LineNumber]="${DC[LineOther]}"
 
         ((++LineNumber))
-        CurrentValueOnLine[LineNumber]="*** ${APP_ENV_FOLDER}/${appname}.env ***"
-        LineColor[LineNumber]="${ColorLineHeading}"
+        CurrentValueOnLine[LineNumber]="*** $(run_script 'app_env_file' "${appname}") ***"
+        LineColor[LineNumber]="${DC[LineHeading]}"
         run_script 'appvars_lines' "${appname}:" > "${CurrentAppEnvFile}"
         local -a CurrentAppEnvLines
         readarray -t CurrentAppEnvLines < <(
@@ -120,23 +104,23 @@ menu_app_vars() {
             VarName="$(grep -o -P '^\w+(?=)' <<< "${line}")"
             if [[ -n ${VarName-} ]]; then
                 # Line contains a variable
-                LineColor[LineNumber]="${ColorLineVar}"
+                LineColor[LineNumber]="${DC[LineVar]}"
                 VarNameOnLine[LineNumber]="${appname}:${VarName}"
                 if [[ -z ${FirstVarLine-} ]]; then
                     FirstVarLine=${LineNumber}
                 fi
             elif (grep -q -P '^\s*#' <<< "${line}"); then
                 # Line is a comment
-                LineColor[LineNumber]="${ColorLineComment}"
+                LineColor[LineNumber]="${DC[LineComment]}"
             else
                 # Line is an unknowwn line
-                LineColor[LineNumber]="${ColorLineOther}"
+                LineColor[LineNumber]="${DC[LineOther]}"
             fi
         done
         ((LineNumber++))
         local AddAppEnvVariableLineNumber=${LineNumber}
         CurrentValueOnLine[LineNumber]="${AddVariableText}"
-        LineColor[LineNumber]="${ColorLineAddVariable}"
+        LineColor[LineNumber]="${DC[LineAddVariable]}"
 
         local TotalLines=$((10#${LineNumber}))
         local PadSize=${#TotalLines}
@@ -155,7 +139,7 @@ menu_app_vars() {
             --ok-label "Select"
             --cancel-label "Done"
             --title "${Title}"
-            --menu "\nApplication: ${ColorHeading}${AppName}\Zn\n" 0 0 0
+            --menu "\n${AppNameHeading}" 0 0 0
             "${LineOptions[@]}"
         )
         while true; do
@@ -192,11 +176,14 @@ menu_app_vars() {
             esac
         done
     done
-    rm -f "${CurrentGlobalEnvFile}" ||
-        warn "Failed to remove temporary .env file.\nFailing command: ${F[C]}rm -f \"${CurrentGlobalEnvFile}\""
-    rm -f "${CurrentAppEnvFile}" ||
-        warn "Failed to remove temporary ${appname}.env file.\nFailing command: ${F[C]}rm -f \"${CurrentAppEnvFile}\""
-
+    if [[ -n ${CurrentGlobalEnvFile} ]]; then
+        rm -f "${CurrentGlobalEnvFile}" ||
+            warn "Failed to remove temporary .env file.\nFailing command: ${F[C]}rm -f \"${CurrentGlobalEnvFile}\""
+    fi
+    if [[ -n ${CurrentAppEnvFile} ]]; then
+        rm -f "${CurrentAppEnvFile}" ||
+            warn "Failed to remove temporary ${appname}.env file.\nFailing command: ${F[C]}rm -f \"${CurrentAppEnvFile}\""
+    fi
 }
 
 test_menu_app_vars() {
